Changed order amount type

+++ Domain/Users/ValueObjects/Order.cs
@@ -4,20 +4,20 @@
     {
         private Order() { }
 
-        private Order(Guid id, Guid userId, decimal amount)
+        private Order(Guid id, Guid userId, int amount)
         {
             Id = id;
             UserId = userId;
             Amount = amount;
         }
 
-        public static Order Create(Guid userId, decimal amount)
+        public static Order Create(Guid userId, int amount)
         {
             return new(Guid.NewGuid(), userId, amount);
         }
 
         public Guid Id { get; private set; }
         public Guid UserId { get; private set; }
-        public decimal Amount { get; private set; }
+        public int Amount { get; private set; }
     }
 }
